+++ extensions/raycast-wallpaper/src/utils/constants.ts
@@ -2,5 +2,5 @@ export const RAYCAST_WALLPAPER_LIST_URL = "https://www.raycast.com/wallpapers.js
 export const RAYCAST_WALLPAPER = "https://www.raycast.com/wallpapers";
 
 export enum LocalStorageKey {
-  WALLPAPER_LIST_CACHE = "Raycast_Wallpapers_List_Cache",
+  WALLPAPER_LIST_CACHE = "Raycast_Wallpaper_List_Cache",
 }