+++ extensions/raycast-wallpaper/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Raycast Wallpaper Changelog
 
+## [Fix bug] - {PR_MERGE_DATE}
+
+- Fix error "AxiosError: Request failed with status code 500"
+
 ## [Add automatic wallpaper switching function] - 2022-08-12
 
 - Change command [Random Raycast Wallpaper] to [Auto Switch Raycast Wallpaper] that will automatically switch random wallpaper every 1 hour when you call it first